+++ src/embedded/label_web.rs
@@ -1,0 +1,7 @@
+use lib_sr::errors::*;
+
+use crate::embedded::html;
+
+pub fn run() -> Result<()> {
+    html::run_with_html(String::from(include_str!("label_web.html")), None)
+}