+++ src/embedded/mod.rs
@@ -17,6 +17,7 @@ use lib_sr::{common, event, Config};
 pub mod generator;
 pub mod html;
 pub mod label;
+pub mod label_web;
 pub mod remove_reviewed;
 pub mod run_using;
 pub mod sink;