+++ src/main.rs
@@ -80,6 +80,8 @@ enum EmbeddedSteps {
     },
     /// Label documents
     Label {},
+    /// Label documents using a web interface
+    LabelWeb {},
     /// Remove documents that have already been reviewed
     RemoveReviewed {},
     /// Run a step using a Nix flake
@@ -109,6 +111,7 @@ fn run_embedded_step(name: EmbeddedSteps) -> Result<()> {
         EmbeddedSteps::Generator { file_or_url } => embedded::generator::run(&file_or_url),
         EmbeddedSteps::Html { file_or_url } => embedded::html::run(&file_or_url),
         EmbeddedSteps::Label {} => embedded::label::run(),
+        EmbeddedSteps::LabelWeb {} => embedded::label_web::run(),
         EmbeddedSteps::RemoveReviewed {} => embedded::remove_reviewed::run(),
         EmbeddedSteps::RunUsing { uses } => embedded::run_using::run(&uses),
         EmbeddedSteps::Sink {} => embedded::sink::run(),